Misc fixes
Unescaped regex grep without delimiters would match 10.0.0.2 when
the local IP was 10.0.0.20.

Errors and undefined variables should stop the script with an error.

Not finding a match should result in and error that makes sense.

Protocol and localhost listener should be able to be overridden.
The defaults from the previous version are preserved.

Global shell variables should be in uppercase.  Shell variables
should be quoted whenever possible.

Add Makefile.